refactor(network): move code around
