--- crates/papyrus_network/src/messages/proto_test/util.proto
@@ -1,17 +1,0 @@
-syntax = "proto3";
-
-message BasicMessage {
-    uint64 number = 1;
-}
-
-message InboundSessionStart {
-    uint64 num_messages = 1;
-    uint64 message_size = 2;
-}
-
-message StressTestMessage {
-    oneof msg {
-        InboundSessionStart start = 1;
-        bytes content = 2;
-    }
-}
