+++ crates/papyrus_network/src/protobuf_messages/proto/p2p/proto/header.proto
