+++ crates/papyrus_network/src/streamed_bytes/behaviour.rs
