+++ crates/papyrus_network/src/streamed_bytes/handler.rs
