+++ crates/papyrus_network/src/streamed_bytes/mod.rs
