+++ crates/papyrus_network/src/streamed_bytes/protocol.rs
